+++ R/paths.R
@@ -1,4 +1,4 @@
-#' Expand Paths of a Formula
+#' Expand paths of a formula
 #'
 #' This presumes that all covariates are confounders are potential confounders.
 #'
@@ -38,39 +38,3 @@ expand_paths <- function(formula, ...) {
 
 }
 
-#' Map Out A Directed Acyclic Graph
-#'
-#' @description
-#' `r lifecycle::badge('experimental')`
-#'
-#' This function converts a hypothesis into a `dagitty` object (or
-#' `tidy_dagitty` if requested). This can subsequently be passed onto the
-#' [ggdag::ggdag()] function for additional plotting.
-#'
-#' @return `dagitty` or `tidy_dagitty` object
-#'
-#' @param study A `study` object
-#'
-#' @param name The name of a hypothesis added to the study
-#'
-#' @param tidy Defaults to FALSE, thus returning a `dagitty` object. If TRUE,
-#'   then will return a `tidy_dagitty` object.
-#'
-#' @importFrom rlang !!!
-#' @export
-draw <- function(study, name, tidy = FALSE) {
-
-	p <- study$path_map
-	f <- p$formulae
-
-	exp <- unique(p$exposures[p$name == name])
-	out <- unique(p$outcomes[p$name == name])
-
-	if (tidy) {
-		rlang::exec(ggdag::dagify, !!!f, exposure = exp, outcome = out) %>%
-		ggdag::tidy_dagitty(.)
-	} else {
-		rlang::exec(ggdag::dagify, !!!f, exposure = exp, outcome = out)
-	}
-
-}